mapas_aplicacao_v1/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
PyYAML>=6.0.1

numpy>=1.26
pandas>=2.1
scipy>=1.11
matplotlib>=3.8

geopandas>=0.14
shapely>=2.0
pyproj>=3.6
fiona>=1.9
rasterio>=1.3

reportlab>=4.0
